requirements.txt
click==6.7
Flask==1.0.0
itsdangerous==0.24
Jinja2>=2.10.1
MarkupSafe==1.0
Werkzeug>=0.14.1
pymongo[tls,srv,gssapi]==3.6.1
Flask_json_schema==0.0.5
flask-cors==3.0.8
APScheduler>= 3.6.1
pycryptodome >= 3.9.0